create the usecases

+++ domain/src/main/java/com/raouf/domain/model/Rating.kt
@@ -1,6 +1,6 @@
 package com.raouf.domain.model
 
 data class Rating(
-    val rate : Float,
-    val count : Int
+    val rate: Double,
+    val count: Int
 )
